Fix #100973: Node Wrangler: previewing node if hierarchy not active
Node Wrangler's preview_node operator can link the active node to the
output node. If the current node tree is a node group inside the base
tree, it can go up the hierarchy to connect until reaching the output
node.

Until now it used a recursive function to go through each active node
from the base tree. If only a single node editor is open, the user can
only open a node group by activating it. But otherwise, they can
select another node at a lower level in the node group hierachy and
the recursion breaks.

This commit changes the operator to use instead the
`SpaceNodeEditor.path`, which gets each node tree in the hierarchy.
This way, we no longer depend on any node in the group hierarchy being
active or not.

The incorrectly named previous commit 0bc87fda67 was only a refactor
and not a fix for this issue.