Fix #104941: Node Wrangler cannot use both bump and normal
Currently, if both bump and normal maps are selected in Principled
Setup, only one of them will be used. But they can both be used in a
PBR setup by plugging the normal output to the bump input.

See for instance Brecht's answer at
https://blender.stackexchange.com/a/16447.

This commit separates setup of bump and normal so they can both be
used at the same time.