FBX IO: Speed up parsing by multithreading array decompression
Because `zlib.decompress` releases the GIL, the arrays are now
decompressed on separate threads.

Given enough logical CPUs on the current system, decompressing arrays
and parsing the rest of the file is now done simultaneously.

This uses the threading utils recently added in `fbx_utils_threading`.

Aside from .fbx files without any compressed arrays, array decompression
usually takes just under 50% of the parsing duration on average, though
commonly varies between 40% to 60% depending on the contents of the file.

However, multithreading array decompression leads to main thread reading
from the file and waiting for IO more often, so I was only able to get
an average of a 35% reduction in parsing duration. Because it's waiting
on IO, this is likely to vary depending on the file system that is being
read from and the time taken to read from IO is expected to be even
longer in real use cases because the file being read won't have been
accessed recently.

For the smallest files, e.g. a single cube mesh, this can be slightly
slower because starting a new thread takes more time than is gained by
starting that thread.

Parsing fbx files takes around 16% of the total import duration on
average, so the overall import duration would be expected to reduce by
about 5.6% on average. However, from timing imports before and after
this patch, I get an actual average reduction of 3.5%.

Because the main thread spends some time waiting on IO, even systems
with a single CPU can see a tiny speedup from this patch. I get about a
6% reduction in parsing duration in this case, which would be less than
1% of the total import duration.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/104739