Fix: Add Mesh Extra Objects: dual tetrahedron is sometimes flipped
Fix an issue with the solid mode of Add Mesh Extra Objects. The dual
tetrahedron gets flipped when it has a vertex truncation of 2.0. It
needs to be mirrored in the same way as the non-dual case.

Noticed while looking into !105117.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/105127