Add Mesh Extra Objects: make solids use ngons
The solid mode of Add Mesh Extra Objects creates quad geometry. This
decision comes from a time when ngons were not supported in Blender,
before the switch to bmesh.

This commit removes the conversion step, as it is better to leave it
up to the user to triangulate or quadrangulate their geometry.