Fix: Add Mesh Extra Objects: Inverted normals of solids when vertex positions are inverted
A number of cases of `add_mesh_solid.createSolid` invert the vertex
positions, but this turns the solid inside-out, causing the normals to
be flipped.

This is fixed by also flipping the normals of the faces in these cases,
which is done by reversing the order of the vertices of each face.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/105123