Node Wrangler: make users of GN node groups independant
A GN node group may be used by multiple modifiers. If multiple nodes
inside a nested node group are viewed by different modifiers, the
other users would get disconnected because the group output was
recreated.

Instead, check for other users of this group and create a new socket
if needed. This was already the behavior for Shaders.