glTF Exporter: Base of export extra animation Like:
    animation of custom property on object
    Bone or object modifier animation

These animations are not glTF compliant, so you need to add data as extras or in an extension.
Data are prepared during export (if option is enabled), and you need an hook to write it into the exported file.
Example of hook: https://github.com/julienduroure/gltf-hooks-examples/blob/main/gltf_extra_animation.py