IO: Option to import each 3DS file on a new collection
Following my initial commit (https://projects.blender.org/blender/blender-addons/pulls/105227), I suggest a enhancement to have an option to include each 3DS file on a new collection (the name of the collection = name of 3DS file).

Importing a 3DS file can result in a lot of objects and it's a mess in the outliner and this commit solve this issue.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/105232