glTF importer:  Use layout panels instead of PanelTypes
Convert glTF to use layout panels instead of PanelTypes for import and
export property drawing. This should match the existing layout used by
the file browser.

This allows viewport/outliner drag-n-drop support to work fully and
it enables the soon to be committed use of Collection exporters.

Thanks Jesse for the patch!