AnimAll: make attribute update a handler instead of an operator
An update is needed for meshes whose elements were animated with old
attribute formats. This update was done inside an operator until now,
but at this point versioning should happen automatically on version
update, as 4.0 was a breaking version anyway.

The update happens on file load instead of version update because the
add-on may not be enabled when the file is updated.