FBX: Enable the Collection exporter feature
This enables FBX to be used as a Collection exporter[0].

To enable this, two changes were made:
- The `export_scene.fbx` operator was configured on the File Handler
- A `collection` property was added to the export Operator

This `collection` property is akin to the special `filepath` and
`directory` properties in that this is how Blender will communicate with
the addon.

Because FBX already has a `use_active_collection` property, a small
layout change is included to not draw this property when used as a
Collection exporter as that would introduce confusion. It also makes
sense to only draw `use_selection` and `use_visible` while in the file
browser in the same way. Feedback was received about this during
development and indeed these toggles will be skipped during processing
anyhow.

An additional discussion point to consider is if we want this new
`collection` property drawn when using `File`->`Export` to allow the
user to control which exact collection to export as opposed to just the
"Active Collection". This isn't included in the patch currently.

[0] To test, go to `Collection Properties` -> `Exporters` -> `Add` and
select FBX after applying this patch.

Pull Request: https://projects.blender.org/blender/blender-addons/pulls/105273