Copy Global Transform: add camera baking & relative copy-paste
Add two new main functions, that are quite related to each other:

**Fix to Camera:**

Also known as "bake to camera", this operator will ensure selected
objects/bones remain static (relative to the camera) on unkeyed frames.

This is done by generating new keys. These keys will be of type
'Generated' so that it remains clear which keys were manually created,
and which were generated. This way the tool can be re-run to re-generate
the keys.

It operates on the scene frame range, or on the preview range if that is
active.

**Relative Copy-Paste:**

The "Relative" panel has copy/paste buttons that work relative to some
object. That object can also be chosen in the panel. If no object is
chosen, the copy/paste will happen relative to the active scene camera.